+++ src/views/layout.tsx
@@ -5,6 +5,7 @@ import {
   PICKER_SCRIPT,
   CONDITIONAL_REVEAL_SCRIPT,
   FORM_ERROR_CLEAR_SCRIPT,
+  SUBMIT_NUDGE_SCRIPT,
 } from "./scripts";
 
 // Meta carries page-level SEO/social metadata into the Layout.
@@ -39,6 +40,7 @@ const GlobalScripts: FC = () => (
     <script dangerouslySetInnerHTML={{ __html: PILL_FILTER_SCRIPT }} />
     <script dangerouslySetInnerHTML={{ __html: PICKER_SCRIPT }} />
     <script dangerouslySetInnerHTML={{ __html: CONDITIONAL_REVEAL_SCRIPT }} />
+    <script dangerouslySetInnerHTML={{ __html: SUBMIT_NUDGE_SCRIPT }} />
   </>
 );